+++ dropbox_backup/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- markdown-link-check-disable -->
+
 ## [0.5.4](https://github.com/mikevansighem/dropbox-backup/releases/tag/0.5.4) - 2021-12-24
 
 <!-- markdown-link-check-enable -->